+++ modules/client/CHANGELOG.md
@@ -20,6 +20,8 @@ TEMPLATE:
 ### Fixed
 - Reading metadata back now parses the IPFS URI schema
 
+### Changed
+- Removed `tokenAddresses` param from `getBalances()` function
 
 ## 0.15.1-alpha
 On 2022-12-06 11:20:50